% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASTD_analyze.R
\name{BASTD_analyze}
\alias{BASTD_analyze}
\title{BASTD_analyze}
\usage{
BASTD_analyze(data, maximum_go_trial_RT)
}
\arguments{
\item{dataframe}{refers to a dataframe containing participant's performance}
}
\value{
BASTD_analyze will return a dataframe with a single row, containing the performance metrics for all the protocols completed by a given participant.
}
\description{
Batch analysis of stop-signal task data or BASTD is a function used to batch analyze SST data
In order to use this function, you may need to change the column name of your output files
The column names required are: ID, Block, Trial, Stimulus, Signal, Correct, Response, RT, RE, SSD
BASTD_analyze is used by OSARI_analyze and STOPIT_analyze
}
\examples{
See OSARI_analyze or STOPIT_analyze for examples

}
