% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSARI_analyze_all.R
\name{OSARI_analyze_all}
\alias{OSARI_analyze_all}
\title{OSARI_analyze_all}
\usage{
OSARI_analyze_all(working_directory)
}
\arguments{
\item{dataframe}{refers to a dataframe containing participant's performance}
}
\value{
OSARI_analyze_all returns a dataframe with rows representing each participant's analyzed data (using the OSARI_analyze function)
}
\description{
Analyze OSARI performance data for all participants
}
\examples{

}
