% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSARI_visualize.R
\name{OSARI_visualize}
\alias{OSARI_visualize}
\title{OSARI_visualize}
\usage{
OSARI_visualize(data)
}
\arguments{
\item{dataframe}{refers to a dataframe containing participant's performance}
}
\value{
BASTD_analyze will return a dataframe with a single row, containing the performance metrics for all the protocols completed by a given participant.
}
\description{
Analyze OSARI performance data for a single participant
}
\examples{
example_OSARI_data <- "https://raw.githubusercontent.com/HeJasonL/BASTD/master/example-data/OSARI_raw_OSARI_2020_Aug_25_1336.txt"
OSARI_data <- read.csv(example_OSARI_data, header = TRUE, sep = "\t")
OSARI_visualize(OSARI_data) #OSARI visualize

}
