% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOPIT_analyze.R
\name{STOPIT_analyze}
\alias{STOPIT_analyze}
\title{STOPIT_analyze}
\usage{
STOPIT_analyze(data)
}
\arguments{
\item{data}{refers to a dataframe containing participant's performance}
}
\value{
STOPIT_analyze will return a dataframe with a single row, containing the performance metrics for all the protocols completed by a given participant.
}
\description{
Analyze OSARI performance data for a single participant
}
\examples{
example_STOP_IT_data <- "https://raw.githubusercontent.com/HeJasonL/BASTD/master/example-data/STOP-IT_raw.csv"
STOP_IT_data <- read.csv(example_STOP_IT_data, header = TRUE) #read the example STOP-IT data
STOPIT_analyze(data = STOP_IT_data) #STOPIT_analyze

}
